修复编译警告: 清除项目代码 39 处 warning (C4477/C4244/C4018/C4302/C4838/C4005)
本次清除项目自身代码的全部可修复警告, 第三方库(skia/mhook)与 MSB8012 工程配置警告不在范围。

C4477 (格式串与参数类型不匹配) 8处:

- SWnd.h 模板 FindChildByID2/FindChildByName2: %s->%ls (LPCWSTR 宽字符参数)

- SObjectFactory.cpp/SResProvider.cpp x2/SSkinPool.cpp: %s->%ls

- demo.cpp: 窄串 %s->%hs, 宽串 %s->%ls

根因: LOGFMT 宏将同一 fmt 同时喂给 _snprintf_s(窄) 与 _snwprintf_s(宽), %s 在两分支语义不同; %ls 在两版均指 wchar_t*, 两分支都匹配。

C4244 (double/int->float/LONG 隐式转换) 20处:

- sdesktopdock.cpp x7 / smatrixwindow.cpp x7 / SPathView.cpp x2 / SFreeMoveWindow.cpp x2 / SHeaderCtrlEx.h x2: 加 (float)/(LONG) 显式转换

C4018 (有符号/无符号比较) 7处:

- SHeaderCtrlEx.h x2 / SPathView.cpp / SShellNotifyIcon.cpp x4: (int) 转换 GetCount() 返回值

C4302 (指针截断) 2处 - 实为真实缺陷:

- spropertygrid.cpp: MAKEINTRESOURCE(IDC_SIZEWE) 把已是资源指针的值再次当 ID 截断为 WORD, 导致光标资源加载错误; 改为直接传 IDC_SIZEWE

C4838 (UINT->LONG 收缩) 1处: imgdecoder-gdip.cpp 加 (LONG) 转换

C4005 (宏重定义) 1处: SIpcObject/stdafx.h 加 #ifndef 守卫

编码处理: UTF-8 BOM (3个demo文件) 与 UTF-16LE (SHeaderCtrlEx.h, SIpcObject/stdafx.h) 均按原编码写回。

验证: Win32 Debug 全量重建 (111s) 后, 过滤 third-part 与 MSB8012, 项目代码 (SOUI/controls.extend/demo) 警告归零。

+++ controls.extend/SFreeMoveWindow.cpp
@@ -201,8 +201,8 @@ namespace SOUI
 			{
 				m_nDistY = rcParent.bottom - rcWnd.bottom;
 			}
-			GetLayoutParam()->SetSpecifiedSize(Horz,SLayoutSize(rcWnd.Width(),SLayoutSize::px));
-			GetLayoutParam()->SetSpecifiedSize(Vert,SLayoutSize(rcWnd.Height(),SLayoutSize::px));
+			GetLayoutParam()->SetSpecifiedSize(Horz,SLayoutSize((float)rcWnd.Width(),SLayoutSize::px));
+			GetLayoutParam()->SetSpecifiedSize(Vert,SLayoutSize((float)rcWnd.Height(),SLayoutSize::px));
 		}
 	}
